change part of speech of {HeghmoH}; +German definitions
The opinions on the mailing list all seem clear on this, and their arguing is very logic. Saying {SoS HeghmoH} can hardly be interpreted as "fatal mother"; besides of that, I don't remember the source, but it was once confirmed that all verbs that include a suffix in their definition are basically just a verb plus a suffix and should be treated as such. 
It's the english translation that is a bit misleading by including the word "be", but unless Maltz gives us a proof for this word to be used as an adjective, I am sure the best way is to follow known rules and treat is as a verb.