fix a bug with the display of person names
Previously, the annotation "(name)" was added to a definition if it was a person name. However, since adding support for other languages, this caused a problem because there is logic to suppress displaying a definition twice if a definition in another language was the same as the English definition. The altering of the English definition broke this logic, .e.g., "Worf (name) / Worf" would be displayed. This change moves the "(name)" annotation to the entry name, as if "is name" is treated as an attribute (similar to "is slang").