Always allow user to edit the query from search results.
Previously, it was only allowed to edit queries with unknown base part of speech, because (most likely) these were entered by the user. Since editing the query strips the part of speech, searching again with the same query does not (necessarily) return the same result, which feels somewhat inconsistent (e.g., if the search had been an internal search on {quv:n}, the query would be "quv" which would show all entries with "quv" as a substring). However, not all internal searches are tagged with part of speech anyway, so some internal sentence searches allowed the query to be edited and others did not. Since there is an inconsistency either way, allowing the query to always be edited might be more useful to the user.